v0.0.10
Removed Schneetarn M70 Jacket from the Corbequais arsenal.
Removed Schneetarn M70 Pants from the Corbequais arsenal.
Added Schneetarn Coveralls to the Corbequais arsenal.
Fixed a bug with the fire rate of painted M16s
Fixed a bug related to the bayonet on the M14.
Fixed a bug related to the bayonet on the M21.
Added a training helmet to the Corbequais arsenal.